refactor: :fire: Uninstall nodemailer package, boilerplate cleanup, file restructuring
Deleted nextjs boilerplate images, removed utils folder from lib, uninstalled nodemailer to work on email authentication at a later time
