+++ lib/mongodb.js
