+++ package.json
@@ -20,7 +20,6 @@
 		"react-dom": "18.2.0"
 	},
 	"devDependencies": {
-		"nodemailer": "^6.9.1",
 		"sass": "^1.57.1"
 	}
 }
